refactor: extract shared text utilities into plugins/text_utils.py
- Create plugins/text_utils.py with strip_markdown(),
  extract_social_title(), slugify() and shared regex constants
- Refactor excerpt_description.py to import from text_utils
- Refactor social_card_title.py to import from text_utils
- Refactor newsletter_email.py to import from plugins.text_utils
  (adds project root to sys.path for standalone script execution)

Net reduction of ~94 lines of duplicated code.
