+++ plugins/social_card_title.py
@@ -17,29 +17,10 @@ import re
 from pymdownx.slugs import slugify as _pymdownx_slugify
 from mkdocs.plugins import BasePlugin, event_priority
 
-# Matches "Delve N: " at the start of a heading (e.g. "Delve 23: ")
-DELVE_PREFIX = re.compile(r"^Delve\s+\d+:\s*", re.IGNORECASE)
+from .text_utils import extract_social_title
 
 # Default pymdownx slugify instance (lowercase, matches Material default)
 _default_slugify = _pymdownx_slugify(case="lower")
-
-
-def extract_social_title(heading: str) -> str:
-    """Extract a short social card title from an H1 heading.
-
-    Strategy:
-      1. Strip the 'Delve N: ' prefix if present.
-      2. If a comma remains (series subtitle separator), take the text after the last comma.
-      3. Otherwise return the remaining text as-is.
-    """
-    # Strip "Delve N: " prefix
-    title = DELVE_PREFIX.sub("", heading)
-
-    # If there's a comma, take the part after the last comma
-    if "," in title:
-        title = title.rsplit(",", 1)[-1].strip()
-
-    return title.strip()
 
 
 def post_slugify(text: str, sep: str = "-") -> str:
